pyproject.toml
[tool.poetry]
name = "genai-python"
version = "0.1.0"
description = ""
authors = ["Your Name <zeeshannoor7831@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.7"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
